+++ app/src/main/java/uz/behzoddev/restaurantpos/domain/interactors/item/UpdateItemInteractor.kt
@@ -3,5 +3,5 @@ package uz.behzoddev.restaurantpos.domain.interactors.item
 import uz.behzoddev.restaurantpos.data.local.models.FoodItem
 
 interface UpdateItemInteractor {
-    suspend operator fun invoke(foodItem: FoodItem): Long
+    suspend operator fun invoke(foodItem: FoodItem): Int
 }
