+++ app/src/main/java/uz/behzoddev/restaurantpos/domain/interactors/item/UpdateItemInteractorImpl.kt
@@ -7,7 +7,7 @@ import javax.inject.Inject
 class UpdateItemInteractorImpl @Inject constructor(
     private val itemRepository: ItemRepository
 ) : UpdateItemInteractor {
-    override suspend fun invoke(foodItem: FoodItem): Long {
+    override suspend fun invoke(foodItem: FoodItem): Int {
         return itemRepository.updateFoodItem(foodItem)
     }
 }
